+++ examples/basic-store/src/pages/Home/models/title.ts
@@ -1,0 +1,10 @@
+export default {
+  state: {
+    title: 'Home Page'
+  },
+  reducers: {
+    update(prevState, payload) {
+      return { title: payload };
+    },
+  }
+};