+++ packages/create-app-shared/src/enhanceWithRouter.ts
@@ -1,6 +1,14 @@
+import { RouteComponentProps, WithRouterStatics, WithRouterProps } from 'react-router';
 import { isMiniAppPlatform } from './env';
 
-function enhanceWithRouter({ withRouter, createElement }) {
+// ref: https://github.com/DefinitelyTyped/DefinitelyTyped/blob/ca5e189854d26c10bd5e9ba078ab33bd3749af43/types/react-router/index.d.ts#L172
+export interface IWithRouter {
+  <P extends RouteComponentProps<any>, C extends React.ComponentType<P>>(
+    component: C & React.ComponentType<P>,
+  ): React.ComponentClass<Omit<P, keyof RouteComponentProps<any>> & WithRouterProps<C>> & WithRouterStatics<C>;
+}
+
+function enhanceWithRouter({ withRouter, createElement }): IWithRouter {
   if (isMiniAppPlatform) {
     withRouter = function (Component) {
       function Wrapper(props) {