+++ packages/plugin-webpack5/src/builtInDependenies.ts
@@ -1,0 +1,15 @@
+// built-in dependencies
+export default [
+  // runApp
+  'universal-env',
+  'miniapp-renderer',
+  'create-app-shared',
+  'react-app-renderer',
+  // plugin-router
+  'query-string',
+  // plugin-store
+  '@ice/store',
+  // plugin-request
+  '@ahooksjs/use-request',
+  'axios',
+];