+++ packages/create-app-shared/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 0.2.3
+
+- [fix] ts types of export apis
+
 ## 0.2.2
 
 - [feat] compatible with cjs and esm when excute runtime module