+++ packages/plugin-store/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.9.6
+
+- [feat] support reset page store after router switch  
+- [feat] support watch store.ts change
+
 ## 1.9.5
 
 - [chore] publish source code of runtime